Added .eval_new_data() in Variable and Call
This new method works with numeric variables and calls that return a numeric variable.
It does not work (or it should not work) with categoric terms because that information is not being memorized yet.

This commit also includes several changes to the structure of the folders.
As always, I start coding without much planification and I end up modifying the structure of a project on the fly.
Please don't be like me, plan ahead.